# lz4block: Optimize amd64 assembly decoder